requirements.txt
# Core RL and Simulation Libraries
gymnasium>=0.29.0
stable-baselines3>=2.1.0
pybullet>=3.2.5
torch>=2.0.0
numpy>=1.24.0

# Visualization and Monitoring
tensorboard>=2.14.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
pandas>=2.0.0

# Optional but recommended
wandb>=0.15.0  # For experiment tracking (optional)
imageio>=2.31.0  # For video recording
imageio-ffmpeg>=0.4.9  # For video encoding
